Render third scene, perlin noise texture

+++ raytracing-the-next-week/src/rtweekend.zig
@@ -34,6 +34,6 @@ pub fn clamp(x: f64, min: f64, max: f64) f64 {
     return x;
 }
 
-pub fn randomInt(min: i32, max: i32) i32 {
-    return @floatToInt(i32, randomDouble2(@intToFloat(f64, min), @intToFloat(f64, max + 1)));
+pub fn randomInt(min: usize, max: usize) usize {
+    return @floatToInt(usize, randomDouble2(@intToFloat(f64, min), @intToFloat(f64, max + 1)));
 }
